+++ alacritty/src/event.rs
@@ -1500,47 +1500,3 @@ impl EventListener for EventProxy {
         let _ = self.0.send_event(Event::TerminalEvent(event));
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn hint_label_generation() {
-        let mut generator = HintLabels::new("0123", 0.5);
-
-        assert_eq!(generator.next(), vec!['0']);
-        assert_eq!(generator.next(), vec!['1']);
-
-        assert_eq!(generator.next(), vec!['2', '0']);
-        assert_eq!(generator.next(), vec!['2', '1']);
-        assert_eq!(generator.next(), vec!['3', '0']);
-        assert_eq!(generator.next(), vec!['3', '1']);
-
-        assert_eq!(generator.next(), vec!['2', '2', '0']);
-        assert_eq!(generator.next(), vec!['2', '2', '1']);
-        assert_eq!(generator.next(), vec!['2', '3', '0']);
-        assert_eq!(generator.next(), vec!['2', '3', '1']);
-        assert_eq!(generator.next(), vec!['3', '2', '0']);
-        assert_eq!(generator.next(), vec!['3', '2', '1']);
-        assert_eq!(generator.next(), vec!['3', '3', '0']);
-        assert_eq!(generator.next(), vec!['3', '3', '1']);
-
-        assert_eq!(generator.next(), vec!['2', '2', '2', '0']);
-        assert_eq!(generator.next(), vec!['2', '2', '2', '1']);
-        assert_eq!(generator.next(), vec!['2', '2', '3', '0']);
-        assert_eq!(generator.next(), vec!['2', '2', '3', '1']);
-        assert_eq!(generator.next(), vec!['2', '3', '2', '0']);
-        assert_eq!(generator.next(), vec!['2', '3', '2', '1']);
-        assert_eq!(generator.next(), vec!['2', '3', '3', '0']);
-        assert_eq!(generator.next(), vec!['2', '3', '3', '1']);
-        assert_eq!(generator.next(), vec!['3', '2', '2', '0']);
-        assert_eq!(generator.next(), vec!['3', '2', '2', '1']);
-        assert_eq!(generator.next(), vec!['3', '2', '3', '0']);
-        assert_eq!(generator.next(), vec!['3', '2', '3', '1']);
-        assert_eq!(generator.next(), vec!['3', '3', '2', '0']);
-        assert_eq!(generator.next(), vec!['3', '3', '2', '1']);
-        assert_eq!(generator.next(), vec!['3', '3', '3', '0']);
-        assert_eq!(generator.next(), vec!['3', '3', '3', '1']);
-    }
-}
